some predictive logic changes (#338)
* some predictive logic changes

* lowering the predictive fill starting volume

* predictive changes

* reworkign the predictive algo

* testing smth else

* wider constrain

* + theoretical avg pump flow

* handling correction for the popup screen

* latest predictive chnages

* applying a higher pressure corection factor

* 90% pz averaging

* changing the default value for PZ

* calculating flow agains smoothed pressure

* slight changes in predictive curve

* bump low pressure flow

* reworked the predictive ranges

* more range correction for the predictive scales

* reverting back to Havald's og flow calc formula

* added a falling flow exclude to  predictive rules

* testing a theory

* lic

* Add release and fill pre-profiling phases

* Try updating the function to compensate for lower pressures

* Ensure FPC never goes below zero

* Add cpp editor config for correctly auto-formatting curly braces

* Second attempt better

* Some more minor tweaks to the flow curve

* Allow EEPROM versioning to upgrade to latest (#347)

This change introduces a mechanism to preserve older versions of the
settings in code, with a process to upgrade to the current version on
flash.

It also allows for arbitrary manipulation of the variables during
version bumps, for instance, to transform a stored data type or data
units, or to selectively override some stored settings with defaults. I
expect this to help changes such as PZ range.

It makes the assumption that versions always increase. It does not
handle versions 1-3, as I tried to reconstruct them from git history,
but couldn't find a consistent schema for most of them.

* indent shandle

* handling steam stuff differently

* + lcd features

* + basket prefill handling

* Calculate weight directly from clicks

* fixes

* fix system uptime minutes

* Add gaggiuino.ino.cpp to gitignore

* Handle fill configuration

Update gaggiuino.ino

* Remove unnecessary commented out constants
