+++ .editorconfig
@@ -7,6 +7,9 @@ charset = utf-8
 trim_trailing_whitespace = true
 indent_style = space
 indent_size = 2
+cpp_new_line_before_open_brace_namespace = same_line
+cpp_new_line_before_open_brace_function = same_line
+cpp_new_line_before_open_brace_type = same_line
 
 [**.{md,adoc}]
 indent_style = space
