+++ .gitignore
@@ -7,3 +7,4 @@ webserver/src/.theia/
 webserver/src/build
 webserver/src/stack/secrets.h
 webserver/data
+gaggiuino.ino.cpp
