+++ platformio.ini
@@ -175,6 +175,9 @@ build_flags =
 	-DPIN_SERIAL2_TX=PA2
 	-DBEAUTIFY_GRAPH
 	-Wdouble-promotion -Wall
+	-O0
+build_unflags =
+	-Os
 
 [env:blackpill-dfu]
 extends = env:blackpill-core
