+++ src/eeprom_data/eeprom_data.h
@@ -1,0 +1,78 @@
+#ifndef EEPROM_DATA_H
+#define EEPROM_DATA_H
+
+#include <Arduino.h>
+#include "log.h"
+
+/**
+* current data version definition below
+*
+* changing the schema requires bumping version number. this forces a schema update
+* on boards with a lower version after flash. to make this graceful for users:
+* - archive the current version in its corresponding header in legacy/
+* - in that header, make sure to register the upgrade function for
+*   the loader to pick it up
+* - bump up the version in eeprom_metadata.h
+* - make schema changes
+*
+* adding fields to data below shouldn't require changes to legacy upgrade
+* functions - they just won't populate the new field, and it will use
+* default value.
+*
+* removing fields might require deleting assignments in existing legacy
+* functions that reference them. this will pop up as a compile time failure
+*/
+
+/**
+* Version 5:
+* - Introduced steamSetpoint
+*/
+struct eepromValues_t {
+  uint16_t setpoint;
+  uint16_t steamSetPoint;
+  uint16_t offsetTemp;
+  uint16_t hpwr;
+  uint16_t mainDivider;
+  uint16_t brewDivider;
+  uint16_t pressureProfilingStart;
+  uint16_t pressureProfilingFinish;
+  uint16_t pressureProfilingHold;
+  uint16_t pressureProfilingLength;
+  bool     pressureProfilingState;
+  bool     preinfusionState;
+  uint16_t preinfusionSec;
+  uint16_t preinfusionBar;
+  uint16_t preinfusionSoak;
+  uint16_t preinfusionRamp;
+  bool     preinfusionFlowState;
+  float    preinfusionFlowVol;
+  uint16_t preinfusionFlowTime;
+  uint16_t preinfusionFlowSoakTime;
+  uint16_t preinfusionFlowPressureTarget;
+  bool     flowProfileState;
+  float    flowProfileStart;
+  float    flowProfileEnd;
+  uint16_t flowProfilePressureTarget;
+  uint16_t flowProfileCurveSpeed;
+  uint16_t powerLineFrequency;
+  uint16_t lcdSleep;
+  bool     warmupState;
+  bool     homeOnShotFinish;
+  bool     graphBrew;
+  bool     brewDeltaState;
+  bool     switchPhaseOnThreshold;
+  bool     basketPrefill;
+  int      scalesF1;
+  int      scalesF2;
+  float    pumpFlowAtZero;
+  bool     stopOnWeightState;
+  float    shotDose;
+  float    shotStopOnCustomWeight;
+  uint16_t shotPreset;
+};
+
+void eepromInit(void);
+bool eepromWrite(eepromValues_t);
+struct eepromValues_t eepromGetCurrentValues(void);
+
+#endif
