+++ src/functional/predictive_weight.h
@@ -3,7 +3,7 @@
 
 #include "../profiling_phases.h"
 #include "../sensors_state.h"
-#include "../eeprom_data.h"
+#include "../eeprom_data/eeprom_data.h"
 
 class PredictiveWeight {
 private:
@@ -20,7 +20,7 @@ public:
   void update(SensorState& state, CurrentPhase& phase, eepromValues_t cfg) {
     // If at least 50ml have been pumped, there has to be output (unless the water is going to the void)
     // No point going through all the below logic if we hardsetting the predictive scales to start counting
-    if (isForceStarted || outputFlowStarted || state.liquidPumped >= 60.f) {
+    if (isForceStarted || outputFlowStarted || state.liquidPumped >= 55.f) {
       outputFlowStarted = true;
       return;
     }
@@ -31,20 +31,29 @@ public:
     puckResistance = pressure * 1000.f / state.smoothedPumpFlow; // Resistance in mBar * s / g
     resistanceDelta = puckResistance - previousPuckResistance;
 
+    // Through empirical testing it's been observed that ~2 bars is the indicator of the pf headspace being full
+    // as well as there being enough pressure for water to wet the puck enough to start the output
     float pressureTarget = phase.getType() == PHASE_TYPE_PRESSURE ? phase.getTarget() : phase.getRestriction();
     pressureTarget = (pressureTarget == 0.f || pressureTarget > 2.f) ? 2.f : pressureTarget;
     // We need to watch when pressure goes above the PI pressure which is a better indicator of headspace being filled.
     // float preinfusionPressure = cfg.preinfusionFlowState ? cfg.preinfusionFlowPressureTarget : cfg.preinfusionBar;
 
+    // If the pressure or flow are raising too fast dismiss the spike from the output.
+    if (state.isPressureRisingFast || state.isPumpFlowRisingFast || state.isPumpFlowFallingFast) {
+      return;
+    }
     // If flow is too big for given pressure or the delta is changing too quickly we're not there yet
     if (resistanceDelta > 500.f || puckResistance < 1100.f) {
       return;
     }
 
     // If pressure or flow is changing quickly we're not there yet
-    if (state.isPressureRisingFast || state.isPumpFlowRisingFast || state.isPumpFlowFallingFast) {
-      return;
-    }
+    // float piPressureTarget = cfg.preinfusionFlowState ? cfg.preinfusionFlowPressureTarget : cfg.preinfusionBar;
+    // if (pressure < piPressureTarget) {
+    //   if (state.isPressureRisingFast || state.isPumpFlowRisingFast) {
+    //     return;
+    //   }
+    // }
 
     // Pressure has to reach at least half way to pressureTarget
     if (pressure < pressureTarget) {
